Fixing changes in types. No longer common directory. Resolving issues within abi and improving unit test
