+++ abis/listener_test.go
@@ -27,35 +27,35 @@ func TestAbiListener(t *testing.T) {
 		contract string
 		expected string
 	}{
-		/* 		{
-		   			name:     "Dummy Contract",
-		   			contract: tests.ReadContractFileForTest(t, "Dummy").Content,
-		   			expected: `[{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"x","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"y","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"uint256","name":"_x","type":"uint256"},{"internalType":"uint256","name":"_y","type":"uint256"}],"type":"constructor"}]`,
-		   		},
-		   		{
-		   			name:     "ERC20 Token",
-		   			contract: tests.ReadContractFileForTest(t, "ERC20_Token").Content,
-		   			expected: `[{"inputs":[],"outputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"name":"MINTER_ROLE","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"subscriptionAmount","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"subscriptionBalance","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isSubscribed","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"rewards","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address[]","name":"","type":"address[]"}],"name":"rewardedUsers","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"string","name":"name","type":"string"},{"internalType":"string","name":"symbol","type":"string"},{"internalType":"uint256","name":"initialSupply","type":"uint256"},{"internalType":"uint256","name":"_subscriptionAmount","type":"uint256"}],"outputs":[],"name":"initialize","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"pause","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"unpause","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"purchaseSubscription","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"cancelSubscription","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"user","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"getSubscriptionStatus","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"reward","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"user","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"getRewards","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address[]","name":"","type":"address[]"}],"name":"getRewardedUsers","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"from","type":"address"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"_beforeTokenTransfer","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"newAmount","type":"uint256"}],"outputs":[],"name":"updateSubscriptionAmount","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"mint","type":"function","stateMutability":"nonpayable"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"SubscriptionPurchased","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"SubscriptionCanceled","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"UserRewarded","type":"event"}]`,
-		   		},
-		   		{
-		   			name:     "Mappings",
-		   			contract: tests.ReadContractFileForTest(t, "Mappings").Content,
-		   			expected: `[{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"simpleMapping","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"doubleMapping","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"tripleMapping","type":"function","stateMutability":"view"}]`,
-		   		},
-		   		{
-		   			name:     "Structs",
-		   			contract: tests.ReadContractFileForTest(t, "Structs").Content,
-		   			expected: `[{"inputs":[{"internalType":"struct MyStructs.ClasicStruct","name":"ClasicStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]},{"internalType":"struct MyStructs.NestedStruct","name":"NestedStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"},{"internalType":"struct MyStructs.ClasicStruct","name":"myStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]}]},{"internalType":"uint256","name":"Integer","type":"uint256"}],"outputs":[{"internalType":"struct MyStructs.NestedStruct","name":"NestedStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"},{"internalType":"struct MyStructs.ClasicStruct","name":"myStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]}]}],"name":"nestedStructExample","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"struct MyStructs.StructWithArray","name":"StructWithArray","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256[]","name":"array","type":"uint256[]"}]}],"outputs":[],"name":"structWithArrayExample","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"struct MyStructs.StructWithNestedArray","name":"StructWithNestedArray","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"struct MyStructs.ClasicStruct[]","name":"structArray","type":"tuple[]","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]}]}],"outputs":[],"name":"structWithNestedArrayExample","type":"function","stateMutability":"pure"}]`,
-		   		},
-		   		{
-		   			name:     "Enums",
-		   			contract: tests.ReadContractFileForTest(t, "Enums").Content,
-		   			expected: `[{"inputs":[],"outputs":[{"internalType":"enum EnumContract.State","name":"","type":"uint8"}],"name":"state","type":"function","stateMutability":"view"},{"inputs":[],"type":"constructor"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isWaiting","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isReady","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isActive","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[],"name":"makeReady","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"makeActive","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"reset","type":"function","stateMutability":"nonpayable"}]`,
-		   		}, */
+		{
+			name:     "Dummy Contract",
+			contract: tests.ReadContractFileForTest(t, "Dummy").Content,
+			expected: `[{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"x","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"y","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"uint256","name":"_x","type":"uint256"},{"internalType":"uint256","name":"_y","type":"uint256"}],"type":"constructor"}]`,
+		},
+		{
+			name:     "ERC20 Token",
+			contract: tests.ReadContractFileForTest(t, "ERC20_Token").Content,
+			expected: `[{"inputs":[],"outputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"name":"MINTER_ROLE","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"subscriptionAmount","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"subscriptionBalance","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isSubscribed","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"rewards","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address[]","name":"","type":"address[]"}],"name":"rewardedUsers","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"string","name":"name","type":"string"},{"internalType":"string","name":"symbol","type":"string"},{"internalType":"uint256","name":"initialSupply","type":"uint256"},{"internalType":"uint256","name":"_subscriptionAmount","type":"uint256"}],"outputs":[],"name":"initialize","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"pause","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"unpause","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"purchaseSubscription","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"cancelSubscription","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"user","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"getSubscriptionStatus","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"reward","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"user","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"getRewards","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address[]","name":"","type":"address[]"}],"name":"getRewardedUsers","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"from","type":"address"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"_beforeTokenTransfer","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"newAmount","type":"uint256"}],"outputs":[],"name":"updateSubscriptionAmount","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"mint","type":"function","stateMutability":"nonpayable"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"SubscriptionPurchased","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"SubscriptionCanceled","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"user","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"UserRewarded","type":"event"}]`,
+		},
+		{
+			name:     "Mappings",
+			contract: tests.ReadContractFileForTest(t, "Mappings").Content,
+			expected: `[{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"simpleMapping","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"doubleMapping","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"tripleMapping","type":"function","stateMutability":"view"}]`,
+		},
+		{
+			name:     "Structs",
+			contract: tests.ReadContractFileForTest(t, "Structs").Content,
+			expected: `[{"inputs":[{"internalType":"struct MyStructs.ClasicStruct","name":"ClasicStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]},{"internalType":"struct MyStructs.NestedStruct","name":"NestedStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"},{"internalType":"struct MyStructs.ClasicStruct","name":"myStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]}]},{"internalType":"uint256","name":"Integer","type":"uint256"}],"outputs":[{"internalType":"struct MyStructs.NestedStruct","name":"NestedStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"},{"internalType":"struct MyStructs.ClasicStruct","name":"myStruct","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]}]}],"name":"nestedStructExample","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"struct MyStructs.StructWithArray","name":"StructWithArray","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256[]","name":"array","type":"uint256[]"}]}],"outputs":[],"name":"structWithArrayExample","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"struct MyStructs.StructWithNestedArray","name":"StructWithNestedArray","type":"tuple","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"struct MyStructs.ClasicStruct[]","name":"structArray","type":"tuple[]","components":[{"internalType":"uint256","name":"one","type":"uint256"},{"internalType":"uint256","name":"two","type":"uint256"}]}]}],"outputs":[],"name":"structWithNestedArrayExample","type":"function","stateMutability":"pure"}]`,
+		},
+		{
+			name:     "Enums",
+			contract: tests.ReadContractFileForTest(t, "Enums").Content,
+			expected: `[{"inputs":[],"outputs":[{"internalType":"enum EnumContract.State","name":"","type":"uint8"}],"name":"state","type":"function","stateMutability":"view"},{"inputs":[],"type":"constructor"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isWaiting","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isReady","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isActive","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[],"name":"makeReady","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"makeActive","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"reset","type":"function","stateMutability":"nonpayable"}]`,
+		},
 		{
 			name:     "Complex Abi",
 			contract: tests.ReadContractFileForTest(t, "AbiComplex").Content,
-			expected: `[]`,
+			expected: `[{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"_msgSender","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bytes","name":"","type":"bytes"}],"name":"_msgData","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"totalSupply","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"account","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"balanceOf","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"transfer","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"owner","type":"address"},{"internalType":"address","name":"spender","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"allowance","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"approve","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"sender","type":"address"},{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"transferFrom","type":"function","stateMutability":"nonpayable"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"from","type":"address"},{"indexed":true,"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"}],"name":"Transfer","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"owner","type":"address"},{"indexed":true,"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"}],"name":"Approval","type":"event"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"add","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"sub","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"},{"internalType":"string","name":"errorMessage","type":"string"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"sub","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"mul","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"div","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"},{"internalType":"string","name":"errorMessage","type":"string"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"div","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"mod","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"a","type":"uint256"},{"internalType":"uint256","name":"b","type":"uint256"},{"internalType":"string","name":"errorMessage","type":"string"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"mod","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"address","name":"account","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isContract","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"sendValue","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"target","type":"address"},{"internalType":"bytes","name":"data","type":"bytes"}],"outputs":[{"internalType":"bytes","name":"","type":"bytes"}],"name":"functionCall","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"target","type":"address"},{"internalType":"bytes","name":"data","type":"bytes"},{"internalType":"string","name":"errorMessage","type":"string"}],"outputs":[{"internalType":"bytes","name":"","type":"bytes"}],"name":"functionCall","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"target","type":"address"},{"internalType":"bytes","name":"data","type":"bytes"},{"internalType":"uint256","name":"value","type":"uint256"}],"outputs":[{"internalType":"bytes","name":"","type":"bytes"}],"name":"functionCallWithValue","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"target","type":"address"},{"internalType":"bytes","name":"data","type":"bytes"},{"internalType":"uint256","name":"value","type":"uint256"},{"internalType":"string","name":"errorMessage","type":"string"}],"outputs":[{"internalType":"bytes","name":"","type":"bytes"}],"name":"functionCallWithValue","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"target","type":"address"},{"internalType":"bytes","name":"data","type":"bytes"},{"internalType":"uint256","name":"weiValue","type":"uint256"},{"internalType":"string","name":"errorMessage","type":"string"}],"outputs":[{"internalType":"bytes","name":"","type":"bytes"}],"name":"_functionCallWithValue","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"_owner","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"_marketingWallet","type":"function","stateMutability":"view"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"prevOwner","type":"address"},{"indexed":true,"internalType":"address","name":"newOwner","type":"address"}],"name":"OwnershipTransferred","type":"event"},{"inputs":[],"type":"constructor"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"owner","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[],"name":"renounceOwnership","type":"function","stateMutability":"nonpayable"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"token0","type":"address"},{"indexed":true,"internalType":"address","name":"token1","type":"address"},{"internalType":"address","name":"pair","type":"address"},{"internalType":"uint256","name":"","type":"uint256"}],"name":"PairCreated","type":"event"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"feeTo","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"feeToSetter","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"tokenA","type":"address"},{"internalType":"address","name":"tokenB","type":"address"}],"outputs":[{"internalType":"address","name":"pair","type":"address"}],"name":"getPair","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"uint256","name":"","type":"uint256"}],"outputs":[{"internalType":"address","name":"pair","type":"address"}],"name":"allPairs","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"allPairsLength","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"tokenA","type":"address"},{"internalType":"address","name":"tokenB","type":"address"}],"outputs":[{"internalType":"address","name":"pair","type":"address"}],"name":"createPair","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[],"name":"setFeeTo","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[],"name":"setFeeToSetter","type":"function","stateMutability":"nonpayable"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"owner","type":"address"},{"indexed":true,"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"}],"name":"Approval","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"from","type":"address"},{"indexed":true,"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"}],"name":"Transfer","type":"event"},{"inputs":[],"outputs":[{"internalType":"string","name":"","type":"string"}],"name":"name","type":"function","stateMutability":"pure"},{"inputs":[],"outputs":[{"internalType":"string","name":"","type":"string"}],"name":"symbol","type":"function","stateMutability":"pure"},{"inputs":[],"outputs":[{"internalType":"uint8","name":"","type":"uint8"}],"name":"decimals","type":"function","stateMutability":"pure"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"totalSupply","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"owner","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"balanceOf","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"owner","type":"address"},{"internalType":"address","name":"spender","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"allowance","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"approve","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"transfer","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"from","type":"address"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"transferFrom","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"name":"DOMAIN_SEPARATOR","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"name":"PERMIT_TYPEHASH","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"address","name":"owner","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"nonces","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"owner","type":"address"},{"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"value","type":"uint256"},{"internalType":"uint256","name":"deadline","type":"uint256"},{"internalType":"uint8","name":"v","type":"uint8"},{"internalType":"bytes32","name":"r","type":"bytes32"},{"internalType":"bytes32","name":"s","type":"bytes32"}],"outputs":[],"name":"permit","type":"function","stateMutability":"nonpayable"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"sender","type":"address"},{"internalType":"uint256","name":"amount0","type":"uint256"},{"internalType":"uint256","name":"amount1","type":"uint256"},{"indexed":true,"internalType":"address","name":"to","type":"address"}],"name":"Burn","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"sender","type":"address"},{"internalType":"uint256","name":"amount0In","type":"uint256"},{"internalType":"uint256","name":"amount1In","type":"uint256"},{"internalType":"uint256","name":"amount0Out","type":"uint256"},{"internalType":"uint256","name":"amount1Out","type":"uint256"},{"indexed":true,"internalType":"address","name":"to","type":"address"}],"name":"Swap","type":"event"},{"anonymous":false,"inputs":[{"internalType":"uint112","name":"reserve0","type":"uint112"},{"internalType":"uint112","name":"reserve1","type":"uint112"}],"name":"Sync","type":"event"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"MINIMUM_LIQUIDITY","type":"function","stateMutability":"pure"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"factory","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"token0","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"token1","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint112","name":"reserve0","type":"uint112"},{"internalType":"uint112","name":"reserve1","type":"uint112"},{"internalType":"uint32","name":"blockTimestampLast","type":"uint32"}],"name":"getReserves","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"price0CumulativeLast","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"price1CumulativeLast","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"kLast","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"to","type":"address"}],"outputs":[{"internalType":"uint256","name":"amount0","type":"uint256"},{"internalType":"uint256","name":"amount1","type":"uint256"}],"name":"burn","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amount0Out","type":"uint256"},{"internalType":"uint256","name":"amount1Out","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"bytes","name":"data","type":"bytes"}],"outputs":[],"name":"swap","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"to","type":"address"}],"outputs":[],"name":"skim","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[],"name":"sync","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"}],"outputs":[],"name":"initialize","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"factory","type":"function","stateMutability":"pure"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"WETH","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"address","name":"tokenA","type":"address"},{"internalType":"address","name":"tokenB","type":"address"},{"internalType":"uint256","name":"amountADesired","type":"uint256"},{"internalType":"uint256","name":"amountBDesired","type":"uint256"},{"internalType":"uint256","name":"amountAMin","type":"uint256"},{"internalType":"uint256","name":"amountBMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountA","type":"uint256"},{"internalType":"uint256","name":"amountB","type":"uint256"},{"internalType":"uint256","name":"liquidity","type":"uint256"}],"name":"addLiquidity","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"token","type":"address"},{"internalType":"uint256","name":"amountTokenDesired","type":"uint256"},{"internalType":"uint256","name":"amountTokenMin","type":"uint256"},{"internalType":"uint256","name":"amountETHMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountToken","type":"uint256"},{"internalType":"uint256","name":"amountETH","type":"uint256"},{"internalType":"uint256","name":"liquidity","type":"uint256"}],"name":"addLiquidityETH","type":"function","stateMutability":"payable"},{"inputs":[{"internalType":"address","name":"tokenA","type":"address"},{"internalType":"address","name":"tokenB","type":"address"},{"internalType":"uint256","name":"liquidity","type":"uint256"},{"internalType":"uint256","name":"amountAMin","type":"uint256"},{"internalType":"uint256","name":"amountBMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountA","type":"uint256"},{"internalType":"uint256","name":"amountB","type":"uint256"}],"name":"removeLiquidity","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"token","type":"address"},{"internalType":"uint256","name":"liquidity","type":"uint256"},{"internalType":"uint256","name":"amountTokenMin","type":"uint256"},{"internalType":"uint256","name":"amountETHMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountToken","type":"uint256"},{"internalType":"uint256","name":"amountETH","type":"uint256"}],"name":"removeLiquidityETH","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"tokenA","type":"address"},{"internalType":"address","name":"tokenB","type":"address"},{"internalType":"uint256","name":"liquidity","type":"uint256"},{"internalType":"uint256","name":"amountAMin","type":"uint256"},{"internalType":"uint256","name":"amountBMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"},{"internalType":"bool","name":"approveMax","type":"bool"},{"internalType":"uint8","name":"v","type":"uint8"},{"internalType":"bytes32","name":"r","type":"bytes32"},{"internalType":"bytes32","name":"s","type":"bytes32"}],"outputs":[{"internalType":"uint256","name":"amountA","type":"uint256"},{"internalType":"uint256","name":"amountB","type":"uint256"}],"name":"removeLiquidityWithPermit","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"token","type":"address"},{"internalType":"uint256","name":"liquidity","type":"uint256"},{"internalType":"uint256","name":"amountTokenMin","type":"uint256"},{"internalType":"uint256","name":"amountETHMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"},{"internalType":"bool","name":"approveMax","type":"bool"},{"internalType":"uint8","name":"v","type":"uint8"},{"internalType":"bytes32","name":"r","type":"bytes32"},{"internalType":"bytes32","name":"s","type":"bytes32"}],"outputs":[{"internalType":"uint256","name":"amountToken","type":"uint256"},{"internalType":"uint256","name":"amountETH","type":"uint256"}],"name":"removeLiquidityETHWithPermit","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"uint256","name":"amountOutMin","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"swapExactTokensForTokens","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amountOut","type":"uint256"},{"internalType":"uint256","name":"amountInMax","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"swapTokensForExactTokens","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amountOutMin","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"swapExactETHForTokens","type":"function","stateMutability":"payable"},{"inputs":[{"internalType":"uint256","name":"amountOut","type":"uint256"},{"internalType":"uint256","name":"amountInMax","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"swapTokensForExactETH","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"uint256","name":"amountOutMin","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"swapExactTokensForETH","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amountOut","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"swapETHForExactTokens","type":"function","stateMutability":"payable"},{"inputs":[{"internalType":"uint256","name":"amountA","type":"uint256"},{"internalType":"uint256","name":"reserveA","type":"uint256"},{"internalType":"uint256","name":"reserveB","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountB","type":"uint256"}],"name":"quote","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"uint256","name":"reserveIn","type":"uint256"},{"internalType":"uint256","name":"reserveOut","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountOut","type":"uint256"}],"name":"getAmountOut","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"amountOut","type":"uint256"},{"internalType":"uint256","name":"reserveIn","type":"uint256"},{"internalType":"uint256","name":"reserveOut","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"}],"name":"getAmountIn","type":"function","stateMutability":"pure"},{"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"getAmountsOut","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"uint256","name":"amountOut","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"}],"outputs":[{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"name":"getAmountsIn","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"token","type":"address"},{"internalType":"uint256","name":"liquidity","type":"uint256"},{"internalType":"uint256","name":"amountTokenMin","type":"uint256"},{"internalType":"uint256","name":"amountETHMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"amountETH","type":"uint256"}],"name":"removeLiquidityETHSupportingFeeOnTransferTokens","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"token","type":"address"},{"internalType":"uint256","name":"liquidity","type":"uint256"},{"internalType":"uint256","name":"amountTokenMin","type":"uint256"},{"internalType":"uint256","name":"amountETHMin","type":"uint256"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"},{"internalType":"bool","name":"approveMax","type":"bool"},{"internalType":"uint8","name":"v","type":"uint8"},{"internalType":"bytes32","name":"r","type":"bytes32"},{"internalType":"bytes32","name":"s","type":"bytes32"}],"outputs":[{"internalType":"uint256","name":"amountETH","type":"uint256"}],"name":"removeLiquidityETHWithPermitSupportingFeeOnTransferTokens","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"uint256","name":"amountOutMin","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[],"name":"swapExactTokensForTokensSupportingFeeOnTransferTokens","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"amountOutMin","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[],"name":"swapExactETHForTokensSupportingFeeOnTransferTokens","type":"function","stateMutability":"payable"},{"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"uint256","name":"amountOutMin","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"},{"internalType":"address","name":"to","type":"address"},{"internalType":"uint256","name":"deadline","type":"uint256"}],"outputs":[],"name":"swapExactTokensForETHSupportingFeeOnTransferTokens","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[{"internalType":"string","name":"","type":"string"}],"name":"_name","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"string","name":"","type":"string"}],"name":"_symbol","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint8","name":"","type":"uint8"}],"name":"_decimals","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"marketingWalletAddress","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"teamWalletAddress","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"deadAddress","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_balances","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_allowances","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isExcludedFromFee","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isWalletLimitExempt","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isTxLimitExempt","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"isMarketPair","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_buyLiquidityFee","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_buyMarketingFee","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_buyTeamFee","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_sellLiquidityFee","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_sellMarketingFee","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_sellTeamFee","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_totalTaxIfBuying","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_totalTaxIfSelling","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_liquidityShare","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_marketingShare","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_teamShare","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_totalDistributionShares","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_totalSupply","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_maxTxAmount","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"_walletMax","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"minimumTokensBeforeSwap","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"contract IUniswapV2Router02","name":"","type":"uint8"}],"name":"uniswapV2Router","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"address","name":"","type":"address"}],"name":"uniswapPair","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"inSwapAndLiquify","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"swapAndLiquifyEnabled","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"swapAndLiquifyByLimitOnly","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"checkWalletLimit","type":"function","stateMutability":"view"},{"anonymous":false,"inputs":[{"internalType":"bool","name":"enabled","type":"bool"}],"name":"SwapAndLiquifyEnabledUpdated","type":"event"},{"anonymous":false,"inputs":[{"internalType":"uint256","name":"tokensSwapped","type":"uint256"},{"internalType":"uint256","name":"ethReceived","type":"uint256"},{"internalType":"uint256","name":"tokensIntoLiqudity","type":"uint256"}],"name":"SwapAndLiquify","type":"event"},{"anonymous":false,"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"}],"name":"SwapETHForTokens","type":"event"},{"anonymous":false,"inputs":[{"internalType":"uint256","name":"amountIn","type":"uint256"},{"internalType":"address[]","name":"path","type":"address[]"}],"name":"SwapTokensForETH","type":"event"},{"inputs":[],"type":"constructor"},{"inputs":[],"outputs":[{"internalType":"string","name":"","type":"string"}],"name":"name","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"string","name":"","type":"string"}],"name":"symbol","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint8","name":"","type":"uint8"}],"name":"decimals","type":"function","stateMutability":"view"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"totalSupply","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"account","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"balanceOf","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"owner","type":"address"},{"internalType":"address","name":"spender","type":"address"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"allowance","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"addedValue","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"increaseAllowance","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"subtractedValue","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"decreaseAllowance","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"newMinimumTokensBeforeSwap","type":"uint256"}],"outputs":[],"name":"setMinimumTokensBeforeSwapAmount","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"minimumTokensBeforeSwapAmount","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"approve","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"owner","type":"address"},{"internalType":"address","name":"spender","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"_approve","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"newLiquidityTax","type":"uint256"},{"internalType":"uint256","name":"newMarketingTax","type":"uint256"},{"internalType":"uint256","name":"newTeamTax","type":"uint256"}],"outputs":[],"name":"setBuyTaxes","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"newLiquidityTax","type":"uint256"},{"internalType":"uint256","name":"newMarketingTax","type":"uint256"},{"internalType":"uint256","name":"newTeamTax","type":"uint256"}],"outputs":[],"name":"setSellTaxes","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"holder","type":"address"},{"internalType":"bool","name":"exempt","type":"bool"}],"outputs":[],"name":"setIsTxLimitExempt","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"account","type":"address"},{"internalType":"bool","name":"newValue","type":"bool"}],"outputs":[],"name":"setIsExcludedFromFee","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"maxTxAmount","type":"uint256"}],"outputs":[],"name":"setMaxTxAmount","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"holder","type":"address"},{"internalType":"bool","name":"exempt","type":"bool"}],"outputs":[],"name":"setIsWalletLimitExempt","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"newLimit","type":"uint256"}],"outputs":[],"name":"setWalletLimit","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"bool","name":"_enabled","type":"bool"}],"outputs":[],"name":"setSwapAndLiquifyEnabled","type":"function","stateMutability":"nonpayable"},{"inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"getCirculatingSupply","type":"function","stateMutability":"view"},{"inputs":[{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[],"name":"transferToAddressETH","type":"function","stateMutability":"nonpayable"},{"type":"receive","stateMutability":"payable"},{"inputs":[{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"transfer","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address[]","name":"addresses","type":"address[]"},{"internalType":"uint256[]","name":"amounts","type":"uint256[]"}],"outputs":[],"name":"transferMulti","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"sender","type":"address"},{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"transferFrom","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"sender","type":"address"},{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"_transfer","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"sender","type":"address"},{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"name":"_basicTransfer","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"tAmount","type":"uint256"}],"outputs":[],"name":"swapAndLiquify","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"tokenAmount","type":"uint256"}],"outputs":[],"name":"swapTokensForEth","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"uint256","name":"tokenAmount","type":"uint256"},{"internalType":"uint256","name":"ethAmount","type":"uint256"}],"outputs":[],"name":"addLiquidity","type":"function","stateMutability":"nonpayable"},{"inputs":[{"internalType":"address","name":"sender","type":"address"},{"internalType":"address","name":"recipient","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"takeFee","type":"function","stateMutability":"nonpayable"}]`,
 		},
 	}
 
@@ -84,7 +84,7 @@ func TestAbiListener(t *testing.T) {
 			assert.NotEmpty(t, abi)
 
 			// Assert the parsed contract matches the expected result
-			//assert.Equal(t, testCase.expected, abiJson)
+			assert.Equal(t, testCase.expected, abiJson)
 		})
 	}
 }
