+++ contracts/contract_listener.go
@@ -4,7 +4,6 @@ import (
 	"strings"
 
 	"github.com/antlr4-go/antlr/v4"
-	"github.com/txpull/solgo/common"
 	"github.com/txpull/solgo/parser"
 )
 
@@ -16,14 +15,14 @@ import (
 type ContractListener struct {
 	*parser.BaseSolidityParserListener                        // BaseSolidityParserListener is the base listener from the Solidity parser.
 	parser                             *parser.SolidityParser // parser is the Solidity parser instance.
-	contractInfo                       common.ContractInfo    // contractInfo is the contract information extracted from the listener.
+	contractInfo                       ContractInfo           // contractInfo is the contract information extracted from the listener.
 }
 
 // NewContractListener creates a new ContractListener. It takes a SolidityParser as an argument.
 func NewContractListener(parser *parser.SolidityParser) *ContractListener {
 	return &ContractListener{
 		parser:       parser,
-		contractInfo: common.ContractInfo{},
+		contractInfo: ContractInfo{},
 	}
 }
 
@@ -110,7 +109,7 @@ func (l *ContractListener) GetComments() []string {
 
 // GetInfoForTests returns a map of all information extracted from the contract.
 // This is used for testing purposes only
-func (l *ContractListener) ToStruct() common.ContractInfo {
+func (l *ContractListener) ToStruct() ContractInfo {
 	return l.contractInfo
 }
